requirements.txt
pymongo
schedule
requests
websocket-client
schedule
python-dotenv
croniter
logging
pandas
tradingview-ta
pytz